belimo-device-simulation/requirements.txt
scipy==1.7.1
pandas==1.3.3
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
oauthlib==3.1.1
requests_oauthlib==1.3.0
numpy==1.21.3